+++ main/dupchess.cpp
@@ -5,7 +5,7 @@
 #include <chrono>
 
 
-#include "io_handler.h"
+#include "chess/io_handler.h"
 
 
 int main()
