+++ src/nnue/CMakeLists.txt
@@ -1,0 +1,10 @@
+cmake_minimum_required(VERSION 3.12)
+
+
+file(GLOB HEADER_FILES CONFIGURE_DEPENDS "*.h")
+file(GLOB SRC_FILES CONFIGURE_DEPENDS "*.cpp")
+
+#uncomment this once there are files in this folder
+#add_library(dupchess_nnue STATIC ${SRC_FILES} ${HEADER_FILES})
+
+target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
